# apps/studio-web/openapi/studio-api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "StudioApi | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/operations/{operationId}": {
      "get": {
        "tags": [
          "Releases"
        ],
        "parameters": [
          {
            "name": "operationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/releases/preview": {
      "post": {
        "tags": [
          "Releases"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishPreview"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishBadRequest"
                }
              }
            }
          }
        }
      }
    },
    "/api/releases": {
      "post": {
        "tags": [
          "Releases"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishBadRequest"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishRejected"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishRejected"
                }
              }
            }
          }
        }
      }
    },
    "/api/releases/{id}": {
      "get": {
        "tags": [
          "Releases"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseSummary"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CategoryDraft": {
        "required": [
          "id",
          "name",
          "selectable"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "selectable": {
            "type": "boolean"
          }
        }
      },
...
      "CategorySides": {
        "required": [
          "published",
          "candidate"
        ],
        "type": "object",
        "properties": {
          "published": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/CategoryDraft"
            }
          },
          "candidate": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/CategoryDraft"
            }
          }
        }
      },
...
      "EditorRef": {
        "required": [
          "id",
          "displayName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          }
        }
      },
...
      "EventDraft": {
        "required": [
          "schemaVersion",
          "title",
          "description",
          "categoryId",
          "occurrenceId",
          "slots"
        ],
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "string"
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "categoryId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "occurrenceId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "slots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventSlot"
            }
          }
        }
      },
...
      "EventSides": {
        "required": [
          "published",
          "candidate"
        ],
        "type": "object",
        "properties": {
          "published": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EventDraft"
              }
            ]
          },
          "candidate": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EventDraft"
              }
            ]
          }
        }
      },
      "EventSlot": {
        "required": [
          "slotId",
          "ocDayId",
          "timeMode",
          "fixed",
          "participation",
          "status",
          "cancelNote",
          "venues"
        ],
        "type": "object",
        "properties": {
          "slotId": {
            "type": "string",
            "format": "uuid"
          },
          "ocDayId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "timeMode": {
            "type": [
              "null",
              "string"
            ]
          },
          "fixed": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FixedTime"
              }
            ]
          },
          "participation": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": "string"
          },
          "cancelNote": {
            "type": [
              "null",
              "string"
            ]
          },
          "venues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SlotVenue"
            }
          }
        }
      },
...
      "FixedTime": {
        "required": [
          "start",
          "end"
        ],
        "type": "object",
        "properties": {
          "start": {
            "type": [
              "null",
              "string"
            ]
          },
          "end": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "OccurrenceDraft": {
        "required": [
          "name",
          "sourceNote",
          "days"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "sourceNote": {
            "type": [
              "null",
              "string"
            ]
          },
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OcDayDraft"
            }
          }
        }
      },
...
      "OccurrenceSides": {
        "required": [
          "published",
          "candidate"
        ],
        "type": "object",
        "properties": {
          "published": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OccurrenceDraft"
              }
            ]
          },
          "candidate": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OccurrenceDraft"
              }
            ]
          }
        }
      },
      "OcDayDraft": {
        "required": [
          "id",
          "date",
          "publicStart",
          "publicEnd",
          "status",
          "cancelNote"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "publicStart": {
            "type": [
              "null",
              "string"
            ]
          },
          "publicEnd": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": "string"
          },
          "cancelNote": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "OperationResult": {
        "required": [
          "operationId",
          "status",
          "startedAt",
          "release",
          "detail"
        ],
        "type": "object",
        "properties": {
          "operationId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "release": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ReleaseSummary"
              }
            ]
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "PreviewDay": {
        "required": [
          "id",
          "occurrenceId",
          "date",
          "publicStart",
          "publicEnd",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "occurrenceId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "publicStart": {
            "type": [
              "null",
              "string"
            ]
          },
          "publicEnd": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": "string"
          }
        }
      },
      "PreviewItem": {
        "required": [
          "targetKind",
          "targetId",
          "label",
          "action",
          "revisionId",
          "publication",
          "event",
          "occurrence",
          "categories",
          "spot"
        ],
        "type": "object",
        "properties": {
          "targetKind": {
            "type": "string"
          },
          "targetId": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "revisionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "publication": {
            "$ref": "#/components/schemas/PublicationSummary"
          },
          "event": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EventSides"
              }
            ]
          },
          "occurrence": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OccurrenceSides"
              }
            ]
          },
          "categories": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CategorySides"
              }
            ]
          },
          "spot": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SpotSides"
              }
            ]
          }
        }
      },
      "PreviewReferences": {
        "required": [
          "days",
          "spots",
          "categories"
        ],
        "type": "object",
        "properties": {
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PreviewDay"
            }
          },
          "spots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PreviewSpot"
            }
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryDraft"
            }
          }
        }
      },
      "PreviewRequest": {
        "required": [
          "entries"
        ],
        "type": "object",
        "properties": {
          "entries": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/PublishEntry"
            }
          }
        }
      },
      "PreviewSpot": {
        "required": [
          "canonicalId",
          "name",
          "buildingName",
          "floorName",
          "utilization"
        ],
        "type": "object",
        "properties": {
          "canonicalId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "buildingName": {
            "type": [
              "null",
              "string"
            ]
          },
          "floorName": {
            "type": [
              "null",
              "string"
            ]
          },
          "utilization": {
            "type": "string"
          }
        }
      },
      "PublicationSummary": {
        "required": [
          "state",
          "publishedRevisionId",
          "publishedAt"
        ],
        "type": "object",
        "properties": {
          "state": {
            "type": "string"
          },
          "publishedRevisionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "publishedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "PublishBadRequest": {
        "required": [
          "code",
          "title"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "PublishEntry": {
        "required": [
          "targetKind",
          "targetId",
          "action",
          "revisionId"
        ],
        "type": "object",
        "properties": {
          "targetKind": {
            "type": "string"
          },
          "targetId": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "revisionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
      "PublishPreview": {
        "required": [
          "validation",
          "items",
          "published",
          "candidate"
        ],
        "type": "object",
        "properties": {
          "validation": {
            "$ref": "#/components/schemas/ValidationRunResult"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PreviewItem"
            }
          },
          "published": {
            "$ref": "#/components/schemas/PreviewReferences"
          },
          "candidate": {
            "$ref": "#/components/schemas/PreviewReferences"
          }
        }
      },
      "PublishRejected": {
        "required": [
          "code",
          "title",
          "validation"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "validation": {
            "$ref": "#/components/schemas/ValidationRunResult"
          }
        }
      },
      "PublishRequest": {
        "required": [
          "operationId",
          "entries",
          "confirmedValidationId",
          "message"
        ],
        "type": "object",
        "properties": {
          "operationId": {
            "type": "string",
            "format": "uuid"
          },
          "entries": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/PublishEntry"
            }
          },
          "confirmedValidationId": {
            "type": "string",
            "format": "uuid"
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ReleaseEntrySummary": {
        "required": [
          "targetKind",
          "targetId",
          "action",
          "revisionId",
          "previousRevisionId",
          "label"
        ],
        "type": "object",
        "properties": {
          "targetKind": {
            "type": "string"
          },
          "targetId": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "revisionId": {
            "type": "string",
            "format": "uuid"
          },
          "previousRevisionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "label": {
            "type": "string"
          }
        }
      },
      "ReleaseSummary": {
        "required": [
          "releaseId",
          "sequence",
          "createdAt",
          "createdBy",
          "source",
          "message",
          "entries"
        ],
        "type": "object",
        "properties": {
          "releaseId": {
            "type": "string",
            "format": "uuid"
          },
          "sequence": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EditorRef"
              }
            ]
          },
          "source": {
            "type": "string"
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReleaseEntrySummary"
            }
          }
        }
      },
...
      "SlotVenue": {
        "required": [
          "canonicalSpotId",
          "note"
        ],
        "type": "object",
        "properties": {
          "canonicalSpotId": {
            "type": "string"
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SpotDraft": {
        "required": [
          "name",
          "aliases",
          "utilization"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "aliases": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "utilization": {
            "type": "string"
          }
        }
      },
...
      "SpotSides": {
        "required": [
          "published",
          "candidate"
        ],
        "type": "object",
        "properties": {
          "published": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SpotDraft"
              }
            ]
          },
          "candidate": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SpotDraft"
              }
            ]
          }
        }
      },
...
      "ValidationFinding": {
        "required": [
          "severity",
          "code",
          "message",
          "targetKind",
          "targetId",
          "targetLabel",
          "path"
        ],
        "type": "object",
        "properties": {
          "severity": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "targetKind": {
            "type": "string"
          },
          "targetId": {
            "type": "string"
          },
          "targetLabel": {
            "type": "string"
          },
          "path": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ValidationRunResult": {
        "required": [
          "id",
          "createdAt",
          "status",
          "blockingCount",
          "warningCount",
          "entries",
          "findings"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "blockingCount": {
            "type": "integer",
            "format": "int32"
          },
          "warningCount": {
            "type": "integer",
            "format": "int32"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublishEntry"
            }
          },
          "findings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationFinding"
            }
          }
        }
      }
    }
  },
...
}